Improve support for providing a different install PREFIX
The Makefile had PREFIX at the top, but any non-standard location
required setting and exporting PKG_CONFIG_PATH so pkgconfig could find
log4cpp.  Add the target bin directory to the PATH to find the
wx-config if built as part of the make.

Even then the script was set to always run sudo, but if PREFIX is
writable by the current user there isn't a need for sudo.